+++ tests/Braintree/CreditCardNumbers/CardTypeIndicators.php
@@ -5,6 +5,7 @@ namespace Test\Braintree\CreditCardNumbers;
 class CardTypeIndicators
 {
     const PREPAID             = "4111111111111210";
+    const PREPAID_RELOADABLE  = "4229989900000002";
     const COMMERCIAL          = "4111111111131010";
     const PAYROLL             = "4111111114101010";
     const HEALTHCARE          = "4111111510101010";